# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger for the API that was developed as a core platform of the BUILDSPACE project.\n**Note** that the **X-Group-Id** is a header that is required in all the enpoints.",
        "title": "BUILSPACE Core Platform Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "BUILDSPACE Core Platform Support",
            "url": "http://www.swagger.io/support",
            "email": "isotiropoulos@singularlogic.eu"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/folder": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a folders meta data by the ID. Pass the ID in a query parameter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Get folder by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a folders meta data by the ID. Pass the Folder model with the updates that are needed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Update folder by ID.",
                "parameters": [
                    {
                        "description": "Update body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Folder"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Use a Folder model as a payload to create a new folder. Essential fields are meta.title (folder's name) and parent (location).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Create a new folder.",
                "parameters": [
                    {
                        "description": "Folder payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Folder"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Bucket"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
        "/folder/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get lists of files and folders in a specific folder, by id. Result is a FolderList model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "List folder's items.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FolderList"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
        "/folder/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pass folder's id to delete it. Nested items (either files or folders) will be deleted as well.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Folders"
                ],
                "summary": "Delete folder by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder payload",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "X-Group-Id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Folder"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorReport"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Bucket": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "models.ErrorReport": {
            "type": "object",
            "properties": {
                "internal_status": {
                    "description": "Status of the error",
                    "type": "string"
                },
                "message": {
                    "description": "Message of the error",
                    "type": "string"
                },
                "reason": {
                    "description": "Reason of the error",
                    "type": "string"
                },
                "status": {
                    "description": "Status of the error",
                    "type": "integer"
                }
            }
        },
...
        "models.Folder": {
            "type": "object",
            "properties": {
                "_id": {
                    "description": "Folder's id",
                    "type": "string"
                },
                "ancestors": {
                    "description": "Array of ancestors' ids",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "files": {
                    "description": "Array of files' ids included",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "folders": {
                    "description": "Array of folders' ids included",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meta": {
                    "description": "Folder's Metadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Meta"
                        }
                    ]
                },
                "parent": {
                    "description": "Parent's folder id",
                    "type": "string"
                }
            }
        },
        "models.FolderList": {
            "type": "object",
            "properties": {
                "files": {
                    "description": "Keys are file ids and values are the files' metadata",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.Meta"
                    }
                },
                "folders": {
                    "description": "Keys are folder ids and values are the folders' metadata",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.Meta"
                    }
                }
            }
        },
        "models.Meta": {
            "type": "object",
            "properties": {
                "creator": {
                    "description": "User's ID that created the file",
                    "type": "string"
                },
                "date_creation": {
                    "description": "Date and time of creation",
                    "type": "integer"
                },
                "descriptions": {
                    "description": "Array of descriptions for the file",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "read": {
                    "description": "Array of user ids with reading rights",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags": {
                    "description": "Array of tags for the file",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title of the file",
                    "type": "string"
                },
                "update": {
                    "description": "Array with data that store the updates",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Updated"
                    }
                },
                "write": {
                    "description": "Array of user ids with writing rights",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Updated": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date and time of update",
                    "type": "string"
                },
                "user": {
                    "description": "User's id that updated",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}